GEODE-10466: Add SerializableRegionRedundancyStatusImpl to serialization allowlist
SerializableRegionRedundancyStatusImpl needs to be added to the sanctioned
serializables allowlist to enable cross-locator serialization in RestoreRedundancy
operations. This class is used to serialize region redundancy status results
when reading them from different locators in DUnit tests.

The serialization filter was correctly rejecting the class because it wasn't
on the allowlist. Adding it enables proper cross-JVM communication for
RestoreRedundancy operations while maintaining the security benefits of the
serialization filter.